Change the condition to flush() the 'kvs::leveldb::WriteBatch'
It flushed only when the buffer if the dirty data was same to
'max_write_queries'.
If the dirty data was greater than 'max_write_queries' (it is a bug) the
buffer will not flushed forever.

It should flush when the dirty data is greater than or equals to
'max_write_queries' to reduce the bug problem for just in case.